Create needleman-wunsch-simple.py
Add an implementation of the Needleman-Wunsch global alignment algorithm using a basic scoring system. A more advanced implementation may use a similarity matrix or scaling gap penalty instead of constant values for matches, mismatches, and gaps as I've done here. Upgrading this code to a more advanced implementation shouldn't be too hard if/when I come back to it.